[DROOLS-6036] move conf impl from drools-core to -compiler
