+++ drools-compiler/src/main/java/org/drools/compiler/builder/impl/JaxbConfigurationFactoryServiceImpl.java
@@ -15,7 +15,7 @@
 package org.drools.compiler.builder.impl;
 
 import com.sun.tools.xjc.Options;
-import org.drools.core.builder.conf.impl.JaxbConfigurationImpl;
+import org.drools.compiler.builder.conf.JaxbConfigurationImpl;
 import org.kie.internal.builder.JaxbConfiguration;
 import org.kie.internal.builder.JaxbConfigurationFactoryService;
