+++ drools-test-coverage/test-suite/src/test/java/org/drools/testcoverage/common/util/ResourceUtil.java
@@ -16,7 +16,7 @@
 
 package org.drools.testcoverage.common.util;
 
-import org.drools.core.builder.conf.impl.DecisionTableConfigurationImpl;
+import org.drools.compiler.builder.conf.DecisionTableConfigurationImpl;
 import org.kie.api.KieServices;
 import org.kie.api.io.Resource;
 import org.kie.internal.builder.DecisionTableConfiguration;
